Rewrite CoreMark port which now compiles
CoreMark port was rewritten to use the barbones example as a baseline rather than Imperas' code.
Implemented a simple Aspire HAL for getting the time using rdcycle.
The rest of the low level CRT stuff comes from the Berkeley riscv-test repo.
This is used to implement the CRT, stdlib stuff like memcpy, and the linker script.

This all works, but I'm questioning if we should go back and use Newlib instead which ships its own
CRT and stdlib functions. This may be more optimal and easier. But in order to do that, we need the
ability to load specs, which only GCC supports.